Added check for redhat and other rpm based distributions to install dependencies.

+++ setup-ansible-node/scripts/install-dependencies.sh
@@ -1,4 +1,9 @@
 #!/usr/bin/env bash
 
 sudo yum install expect -y
-sudo yum install epel-release -y
+
+if [ ! -f /etc/redhat-release ]; then
+  sudo yum install epel-release -y
+fi
+
+
